refac(fix/input): 생일 다이얼에서 인풋으로 변경
