+++ components/profile-modify/ModifyBirth.vue
@@ -4,54 +4,15 @@
             <span class="modify-birth-title">생년월일</span>
         </div>
         <div class="modify-birth-content">
-            <div class="year-scroller">
-                <div
-                    class="year"
-                    v-for="year in displayedYears"
-                    :key="'year-' + year"
-                    :class="{ 'current-year': year === currentYear }"
-                    @touchstart="handleDragStart"
-                    @touchmove="handleDragMove"
-                    @touchend="handleDragEnd"
-                >
-                    {{ year }}
-                </div>
-            </div>
-            <div class="year-text-div">
-                <span class="year-text">년</span>
-            </div>
-            <div class="month-scroller">
-                <div
-                    class="month"
-                    v-for="month in displayedMonths"
-                    :key="'month-' + month"
-                    :class="{ 'current-month': month === currentMonth }"
-                    @touchstart="handleDragStart"
-                    @touchmove="handleDragMove"
-                    @touchend="handleDragEnd"
-                >
-                    {{ month }}
-                </div>
-            </div>
-            <div class="month-text-div">
-                <span class="month-text">월</span>
-            </div>
-            <div class="day-scroller">
-                <div
-                    class="day"
-                    v-for="day in displayedDays"
-                    :key="'day-' + day"
-                    :class="{ 'current-day': day === currentDay }"
-                    @touchstart="handleDragStart"
-                    @touchmove="handleDragMove"
-                    @touchend="handleDragEnd"
-                >
-                    {{ day }}
-                </div>
-            </div>
-            <div class="day-text-div">
-                <span class="day-text">일</span>
-            </div>
+            <MutateCalendar
+                :time="currentBirth"
+                :dateErrMsgObj="dateErrMsgObj.currentBirth"
+                :dateErrorMsg="dateErrorMsg.currentBirth"
+                @validateYear="validateYear"
+                @validateMonth="validateMonth"
+                @validateDay="validateDay"
+                :isBirthComponent="true"
+            />
         </div>
     </div>
 </template>
@@ -59,189 +20,121 @@
 <script>
 import { mapState } from "pinia";
 import { useUserStore } from "~/store/user";
+import MutateCalendar from "~/components/calendar/mutate/MutateCalendar.vue";
 
 export default {
     name: "ModifyBirth",
+    components: {
+        MutateCalendar,
+    },
     data() {
         return {
-            currentYear: 2000,
-            displayedYears: [],
-            currentMonth: 1,
-            displayedMonths: [],
-            currentDay: 1,
-            displayedDays: [],
-            startY: 0,
-            lastY: 0,
-            startX: 0,
-            lastX: 0,
-            lastTime: Date.now(),
-            velocity: 0,
-            touchData: {
-                startY: 0,
-                lastY: 0,
-                lastTime: Date.now(),
-                velocity: 0,
+            currentBirth: {
+                year: null,
+                month: null,
+                day: null,
             },
-            scrollingAreas: [
-                { startX: 22, endX: 105, handleScroll: this.handleYearScroll },
-                {
-                    startX: 155,
-                    endX: 225,
-                    handleScroll: this.handleMonthScroll,
-                },
-                { startX: 280, endX: 350, handleScroll: this.handleDayScroll },
-            ],
+            dateErrMsgObj: { currentBirth: {} },
+            dateErrorMsg: { currentBirth: "" },
         };
     },
     computed: {
         ...mapState(useUserStore, ["user"]),
     },
-    mounted() {
+    created() {
         if (this.user?.birth) {
             const [year, month, day] = this.user.birth.split("-");
-            this.currentYear = parseInt(year);
-            this.currentMonth = parseInt(month);
-            this.currentDay = parseInt(day);
+            this.currentBirth.year = parseInt(year);
+            this.currentBirth.month = parseInt(month);
+            this.currentBirth.day = parseInt(day);
         }
-
-        this.setDisplayedYears();
-        this.setDisplayedMonths();
-        this.setDisplayedDays();
     },
     methods: {
-        checkYear() {
-            if (this.currentYear < 1900) this.currentYear = 1900;
-            if (this.currentYear > 2023) this.currentYear = 2023;
-        },
-        checkMonth() {
-            if (this.currentMonth < 1) this.currentMonth = 1;
-            if (this.currentMonth > 12) this.currentMonth = 12;
-        },
-        checkDay() {
-            const maxDaysInMonth = new Date(
-                this.currentYear,
-                this.currentMonth,
-                0,
-            ).getDate();
-            if (this.currentDay < 1) this.currentDay = 1;
-            if (this.currentDay > maxDaysInMonth)
-                this.currentDay = maxDaysInMonth;
-        },
-        setDisplayedYears() {
-            this.checkYear();
-            this.displayedYears = [
-                this.currentYear - 1,
-                this.currentYear,
-                this.currentYear + 1,
-            ];
-        },
-        setDisplayedMonths() {
-            this.checkMonth();
-            let previousMonth = this.currentMonth - 1;
-            let nextMonth = this.currentMonth + 1;
-            this.displayedMonths = [
-                previousMonth || 12,
-                this.currentMonth,
-                nextMonth === 13 ? 1 : nextMonth,
-            ];
-        },
-        setDisplayedDays() {
-            this.checkDay();
-            let previousDay = this.currentDay - 1;
-            let nextDay = this.currentDay + 1;
-            this.displayedDays = [
-                previousDay || 31,
-                this.currentDay,
-                nextDay === 32 ? 1 : nextDay,
-            ];
-        },
-        handleDragStart(event) {
-            // event.stopPropagation();
-            event.preventDefault();
-            this.startY = event.touches[0].clientY;
-            this.startX = event.touches[0].clientX;
-            this.touchData.lastY = this.startY;
-            this.touchData.lastTime = Date.now();
-        },
-        handleDragMove(event) {
-            // event.stopPropagation();
-            event.preventDefault();
-            const currentX = event.touches[0].clientX;
-            const currentY = event.touches[0].clientY;
-            let deltaY = currentY - this.touchData.lastY;
-            const sensitivity = 0.05;
-            deltaY *= sensitivity;
+        validateYear(placeToCall) {
+            const year = parseInt(this[placeToCall].year) || 0;
+            this[placeToCall].year = year;
 
-            const deltaTime = Date.now() - this.touchData.lastTime;
-            this.touchData.velocity = deltaY / deltaTime;
-            this.touchData.lastY = currentY;
-            this.touchData.lastTime = Date.now();
+            if (isNaN(year) || year < 1000 || year > 9999) {
+                this.dateErrMsgObj[placeToCall].year =
+                    "올바른 년도를 입력해주세요.";
+                this.emitStopSubmit();
+            } else {
+                this.dateErrMsgObj[placeToCall].year = "";
 
-            for (const area of this.scrollingAreas) {
-                if (currentX >= area.startX && currentX <= area.endX) {
-                    area.handleScroll(deltaY);
-                    break;
+                if (this.isEditMode) {
+                    this.updateIsRemodified(true);
                 }
             }
+            this.getDateErrMsg(placeToCall);
         },
-        handleDragEnd(event) {
-            // event.stopPropagation();
-            this.lastX = event.changedTouches[0].clientX;
-            const inertiaDuration = 100;
-            const distance = this.touchData.velocity * inertiaDuration;
-            const change = distance / 36;
+        validateMonth(placeToCall) {
+            const month = parseInt(this[placeToCall].month) || 0;
+            this[placeToCall].month = month;
 
-            if (this.lastX >= 22 && this.lastX <= 105) {
-                this.currentYear = Math.round(this.currentYear - change);
-                this.setDisplayedYears();
-            } else if (this.lastX >= 155 && this.lastX <= 225) {
-                this.currentMonth = Math.round(this.currentMonth - change);
-                this.setDisplayedMonths();
-            } else if (this.lastX >= 280 && this.lastX <= 350) {
-                this.currentDay = Math.round(this.currentDay - change);
-                this.setDisplayedDays();
+            if (isNaN(month) || month < 1 || month > 12) {
+                this.dateErrMsgObj[placeToCall].month =
+                    "올바른 월을 입력해주세요.";
+                this.emitStopSubmit();
+            } else {
+                this.dateErrMsgObj[placeToCall].month = "";
+
+                if (this.isEditMode) {
+                    this.updateIsRemodified(true);
+                }
+            }
+            this.getDateErrMsg(placeToCall);
+        },
+        validateDay(placeToCall) {
+            const day = parseInt(this[placeToCall].day) || 0;
+            this[placeToCall].day = day;
+
+            if (isNaN(day) || day < 1 || day > 31) {
+                this.dateErrMsgObj[placeToCall].day =
+                    "올바른 일을 입력해주세요.";
+                this.emitStopSubmit();
+            } else {
+                this.dateErrMsgObj[placeToCall].day = "";
+
+                if (this.isEditMode) {
+                    this.updateIsRemodified(true);
+                }
+            }
+            this.getDateErrMsg(placeToCall);
+        },
+        getDateErrMsg(placeToCall) {
+            if (this.dateErrMsgObj[placeToCall].year) {
+                this.dateErrorMsg[placeToCall] =
+                    this.dateErrMsgObj[placeToCall].year;
+            } else if (this.dateErrMsgObj[placeToCall].month) {
+                this.dateErrorMsg[placeToCall] =
+                    this.dateErrMsgObj[placeToCall].month;
+            } else {
+                this.dateErrorMsg[placeToCall] =
+                    this.dateErrMsgObj[placeToCall].day;
+            }
+
+            if (!this.dateErrorMsg[placeToCall]) {
+                this.emitBirthChange(placeToCall);
             }
         },
         emitBirthChange() {
-            const birth = `${this.currentYear}-${String(
-                this.currentMonth,
-            ).padStart(2, "0")}-${String(this.currentDay).padStart(2, "0")}`;
+            if (
+                this.currentBirth.year &&
+                this.currentBirth.month &&
+                this.currentBirth.day
+            ) {
+                this.$emit("onStopSubmit", true);
+            }
+            const birth = `${this.currentBirth.year}-${String(
+                this.currentBirth.month,
+            ).padStart(2, "0")}-${String(this.currentBirth.day).padStart(
+                2,
+                "0",
+            )}`;
             this.$emit("birthSelected", birth);
         },
-        handleYearScroll(deltaY) {
-            const increment = deltaY >= 0.3 ? -1 : deltaY <= -0.3 ? 1 : 0;
-            this.currentYear += increment;
-            this.setDisplayedYears();
-            this.emitBirthChange();
-        },
-        handleMonthScroll(deltaY) {
-            const increment = deltaY >= 0.3 ? -1 : deltaY <= -0.3 ? 1 : 0;
-            this.currentMonth += increment;
-
-            if (this.currentMonth === 0) {
-                this.currentMonth = 12;
-            } else if (this.currentMonth === 13) {
-                this.currentMonth = 1;
-            }
-            this.setDisplayedMonths();
-            this.emitBirthChange();
-        },
-        handleDayScroll(deltaY) {
-            const maxDaysInMonth = new Date(
-                this.currentYear,
-                this.currentMonth,
-                0,
-            ).getDate();
-            const increment = deltaY >= 0.3 ? -1 : deltaY <= -0.3 ? 1 : 0;
-            this.currentDay += increment;
-
-            if (this.currentDay === 0) {
-                this.currentDay = maxDaysInMonth;
-            } else if (this.currentDay > maxDaysInMonth) {
-                this.currentDay = 1;
-            }
-            this.setDisplayedDays();
-            this.emitBirthChange();
+        emitStopSubmit() {
+            this.$emit("onStopSubmit", false);
         },
     },
 };
@@ -270,79 +163,5 @@ export default {
     // height: 144px;
     display: flex;
     justify-content: space-around;
-}
-.year-scroller,
-.month-scroller,
-.day-scroller {
-    // height: 144px;
-    // width: 80px;
-    flex-direction: column;
-    border-radius: 8px;
-    align-items: center;
-    padding-top: 1.3rem;
-}
-
-.year,
-.month,
-.day {
-    // height: 30px;
-    // width: 80px;
-    margin: 6px 0;
-    display: flex;
-    justify-content: center;
-    align-items: center;
-    color: var(--gray-400, #9ca3af);
-    text-align: center;
-    font-family: Pretendard;
-    font-size: 20px;
-    line-height: 150%;
-    &.current-year,
-    &.current-month,
-    &.current-day {
-        // width: 80px;
-        // height: 60px;
-        padding: 1rem;
-        flex-shrink: 0;
-        margin-left: 0;
-        border-radius: 8px;
-        background: var(--indigo-50, #eef2ff);
-        color: var(--indigo-500, #6366f1);
-        text-align: center;
-        font-family: "Pretendard Medium";
-        font-size: 24px;
-        line-height: 150%;
-        display: flex;
-        justify-content: center;
-        align-items: center;
-    }
-    &.current-month,
-    &.current-day {
-        // width: 54px;
-    }
-}
-.day,
-.month {
-    // width: 54px;
-}
-.year-text-div,
-.month-text-div,
-.day-text-div {
-    height: 100%;
-    display: flex;
-    align-items: center;
-}
-.year-text,
-.month-text,
-.day-text {
-    color: var(--gray-500, #6b7280);
-    text-align: center;
-    font-family: "Pretendard Bold";
-    margin-left: 8px;
-    font-size: 18px;
-    line-height: 160%;
-}
-.year-text,
-.month-text {
-    margin-right: 2rem;
 }
 </style>
